+++ README.md
@@ -4,12 +4,17 @@ Default Course Project of CS4278/CS5478 Intelligent Robots: Algorithms and Syste
 ![Scene](imgs/scene.png)
 
 
-In this project, you should accomplish:
+In this project, the task is to:
 
-1. Navigate the robot to the bedside table to see the cup. The goal image is at `imgs/goal.png`.
-2. Pick up the cup.
-3. Open the drawer and place the cup in it, then close the drawer.
+1. Navigate the Mobot to the drawer
+2. Pick up the mug
+3. Place it into the drawer
 
+We provide an example control code for the robot in `simulation/main.py`.
+
+# Requirement
+
+You should implement the navigation and motion planning algorithms by yourself to accomplish the task.
 
 
 # Installation
@@ -28,4 +33,6 @@ In this project, you should accomplish:
 
 # References
 
-PyBullet Documentation: https://pybullet.org/wordpress/index.php/forum-2/
+PyBullet Documentation: https://pybullet.org/wordpress/index.php/forum-2/.
+
+You can also use ChatGPT or Cursor to get the API of PyBullet.
